+++ build/app/Controllers/Http/UserController.js
@@ -12,18 +12,24 @@ const jsonwebtoken_1 = __importDefault(require("jsonwebtoken"));
 class UsersController {
     async index({ response, request }) {
         try {
-            const user = await User_1.default.findOrFail(jsonwebtoken_1.default.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id']);
-            if (user.role.role_id == 5 || user.role.role_id == 4) {
-                const users = await User_1.default.query()
-                    .preload('role')
-                    .preload('credit_card')
-                    .preload('delivery_address_id')
-                    .preload('payment_address_id')
-                    .preload('restorer');
-                return response.status(200).json({ users });
+            const user_id = this.getId(request);
+            if (user_id) {
+                const user = await User_1.default.findOrFail(user_id);
+                if (user.role.role_id == 5 || user.role.role_id == 4) {
+                    const users = await User_1.default.query()
+                        .preload('role')
+                        .preload('credit_card')
+                        .preload('delivery_address_id')
+                        .preload('payment_address_id')
+                        .preload('restorer');
+                    return response.status(200).json({ users });
+                }
+                else {
+                    return response.status(403).json({ message: "Error you don't have the correct permissions" });
+                }
             }
             else {
-                return response.status(403).json({ message: "Error you don't have the correct permissions" });
+                return response.status(500).json({ err: "jwt token error" });
             }
         }
         catch (err) {
@@ -32,8 +38,16 @@ class UsersController {
     }
     async getById({ response, request }) {
         try {
-            const user = await User_1.default.findOrFail(jsonwebtoken_1.default.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id']);
-            return response.status(200).json({ user });
+            const user_id = this.getId(request);
+            console.log(user_id);
+            if (user_id) {
+                const user = await User_1.default.findBy('user_id', user_id);
+                console.log(user);
+                return response.status(200).json({ user });
+            }
+            else {
+                return response.status(500).json({ err: "jwt token error" });
+            }
         }
         catch (err) {
             return response.status(500).json({ err });
@@ -108,10 +122,16 @@ class UsersController {
     }
     async update({ request, response }) {
         try {
-            const user = await User_1.default.findOrFail(jsonwebtoken_1.default.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id']);
-            user.merge({ user_firstname: request.body()['user_firstname'], user_lastname: request.body()['user_lastname'], user_email: request.body()['user_email'], password: request.body()['user_password'], user_phone_number: request.body()['user_phone_number'] });
-            await user.save();
-            return response.status(200).json({ user });
+            const user_id = this.getId(request);
+            if (user_id) {
+                const user = await User_1.default.findOrFail(user_id);
+                user.merge({ user_firstname: request.body()['user_firstname'], user_lastname: request.body()['user_lastname'], user_email: request.body()['user_email'], password: request.body()['user_password'], user_phone_number: request.body()['user_phone_number'] });
+                await user.save();
+                return response.status(200).json({ user });
+            }
+            else {
+                return response.status(500).json({ err: "jwt token error" });
+            }
         }
         catch (err) {
             return response.status(500).json({ err });
@@ -119,22 +139,28 @@ class UsersController {
     }
     async updateSponsor({ request, response }) {
         try {
-            const user = await User_1.default.findOrFail(jsonwebtoken_1.default.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id']);
-            const filleul = await User_1.default.findBy('user_email', request.body()['filleul_email']);
-            if (user.user_support == false && filleul && user.fk_role_id == filleul.fk_role_id) {
-                if (filleul.user_is_supported == false) {
-                    user.user_support = true;
-                    await user.save();
-                    filleul.user_is_supported = true;
-                    await filleul.save();
+            const user_id = this.getId(request);
+            if (user_id) {
+                const user = await User_1.default.findOrFail(user_id);
+                const filleul = await User_1.default.findBy('user_email', request.body()['filleul_email']);
+                if (user.user_support == false && filleul && user.fk_role_id == filleul.fk_role_id) {
+                    if (filleul.user_is_supported == false) {
+                        user.user_support = true;
+                        await user.save();
+                        filleul.user_is_supported = true;
+                        await filleul.save();
+                    }
+                    else {
+                        return response.status(400).json({ message: "error this user is already supported or you enter the wrong email" });
+                    }
+                    return response.status(200).json({ user });
                 }
                 else {
-                    return response.status(400).json({ message: "error this user is already supported or you enter the wrong email" });
+                    return response.status(400).json({ message: "error Unable to find the user or the user is already supporting someone" });
                 }
-                return response.status(200).json({ user });
             }
             else {
-                return response.status(400).json({ message: "error Unable to find the user or the user is already supporting someone" });
+                return response.status(500).json({ err: "jwt token error" });
             }
         }
         catch (err) {
@@ -143,12 +169,29 @@ class UsersController {
     }
     async delete({ response, request }) {
         try {
-            const user = await User_1.default.findOrFail(jsonwebtoken_1.default.verify(request.input('jwt'), 'TOKEN_PRIVATE_KEY')['user_id']);
-            await user.delete();
-            return response.status(200).json({ user });
+            const user_id = this.getId(request);
+            if (user_id) {
+                const user = await User_1.default.findOrFail(user_id);
+                await user.delete();
+                return response.status(200).json({ user });
+            }
+            else {
+                return response.status(500).json({ err: "jwt token error" });
+            }
         }
         catch (err) {
             return response.status(500).json({ err: err, message: "wrong user" });
+        }
+    }
+    getId(request) {
+        const token = request.header('authorization')?.split(" ");
+        try {
+            let user_id = jsonwebtoken_1.default.verify(token[1], 'TOKEN_PRIVATE_KEY')['user_id'];
+            console.log(user_id);
+            return user_id;
+        }
+        catch {
+            return false;
         }
     }
 }
